Style heading
I've aligned my main heading content to the center and made it the way it won't virtually grow when you try to zoom out. However, the navigation bar will grow and stay in the corners of the page.
I've added styling to each element, so they look pretty and organically fit with each other.